alarm ringing screen, snooze and dismiss

+++ app/src/main/java/com/volfor/ondori/features/alarm/domain/services/AlarmScheduler.kt
@@ -1,6 +1,6 @@
 package com.volfor.ondori.features.alarm.domain.services
 
 interface AlarmScheduler {
-    fun scheduleAlarm(alarmId: Int, triggerAtMillis: Long)
-    fun cancelAlarm(alarmId: Int)
+    fun scheduleAlarm(alarmId: Long, triggerAtMillis: Long)
+    fun cancelAlarm(alarmId: Long)
 }
